Rake task to set new Publishing API dates
This is used to populate the date fields of:

- temporary_first_published_at
- publisher_first_published_at
- major_published_at
- publisher_major_published_at
- published_at
- publisher_published_at
- temporary_last_edited_at
- publisher_last_edited_at

To be consistent with the new date fields introduced by:
https://github.com/alphagov/publishing-api/pull/1007

This is a slow process, I think the script probably takes ~6 hours to
run full. It also has a few limitations:

- Only able to use events when there is a single locale
- Unable to work out major_publishing when the update_type was set in
publish (lost in events at that point)
- Makes some approximations based on nano-second value

My feeling is that although this isn't completely accurate, it's
probably accurate enough to lay some decent foundation. It introduces
a Edition#new_dates_valid? method which can be used to check that the
dates haven't deviated from what will be presented to Content Store and
other services.